flower_scratch/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flower_scratch"
version = "1.0.0"
description = ""
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.13.1",
    "flwr-datasets[vision]>=0.3.0",
    "torch==2.2.1",
    "torchvision==0.17.1",
    "wandb"
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "kvnptl"

[tool.flwr.app.components]
serverapp = "flower_scratch.server_app:app"
clientapp = "flower_scratch.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 3
fraction-fit = 0.5
local-epochs = 1

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 3

[tool.flwr.federations.larger-sim]
options.num-supernodes = 3
options.backend.client-resources.num-cpus = 1
options.backend.client-resources.num-gpus = 0.5